+++ constants/country.constant.js
@@ -1,0 +1,4 @@
+export const countryStatus = {
+    active: 1,
+    inactive: 0
+}
