+++ constants/role.constant.js
@@ -1,0 +1,4 @@
+export const roleStatus = {
+    active: 1,
+    inactive: 0
+}
